Test change to send different telemetry
