+++ connectors/services/job_scheduling.py
@@ -33,7 +33,7 @@ class JobSchedulingService(BaseService):
     name = "schedule"
 
     def __init__(self, config):
-        super().__init__(config)
+        super().__init__(config, "Job Scheduling Service")
         self.idling = self.service_config["idling"]
         self.heartbeat_interval = self.service_config["heartbeat"]
         self.source_list = config["sources"]
